better error msgs, check seq length,  updated cython to use ctypedef struct defs

+++ c/lib/seq.h
@@ -26,7 +26,7 @@ typedef struct {
 Seq *seq_new(void);
 void seq_expand(Seq *seq);
 void seq_read_str(Seq *seq, char *seq_str);
-int seq_read_fasta_record(Seq *seq, gzFile f);
+long seq_read_fasta_record(Seq *seq, gzFile f);
 void seq_write_fasta_record(Seq *seq, gzFile f);
 void seq_rev(Seq *fwd_seq);
 void seq_comp(Seq *seq);
